Fixed several bugs in the checkpoint code.
Recovering a hidden expression with bracket indices crashed the program, because
the term compression was switched off after a restart and the bracket indices
would point now into different places in the hidden expression. The reason for
the compression switch-off was that AT.SS was not pointing to AT.S0, as it does
at any module beginning following at least one sorting module.

Some temporary files where not removed.

Recovering time information for 1 or 2 worker threads crashed the program and
for 3 or more worker threads it gave wrong timings.

Log files were overridden and not re-opened correctly.

Status messages didn't work with log-files (printf->MesPrint).