Changed the version to 4.0Beta in configure.ac
This commit also includes:
BETAVERSION is defined in config.h when VERSION_IS_BETA in configure.ac
is "yes".
The format of version message at runtime is modified (< 79 chars). Also
the macros FORMNAME and VERSIONSTR are defined in startup.c.
MAJORVERSION and MINORVERSION are reflected to VERSION_ and SUBVERSION_,
respectively.
Fix for ParFORM: only master prints the version.