ParFORM: new statistics report
Now ParFORM can report the statistics in a way similar to TFORM. Two
options have been added, processstats and oldparallelstats, which
control how ParFORM reports the statistics. `On processstats' is similar
to `On threadstats' (default is on). `On oldparallelstats' switches the
statistics mode to the old one (default is off).