ParFORM: small changes in the statistics
The master now reports the number of generated terms as a sum of the
numbers of generated terms in the slaves (instead of a sum of the
numbers of terms sent from the slaves) in the similar way as TFORM.

Unnecessary empty lines in the statistics have been removed.

PF_EndSort has been changed such that it sets S->TermsLeft.
(This change fixed ZERO_expr.) EndSort was simplified a bit.

PF.goutterms and PF.ExprSize have been localised as PF_goutterms and
PF_exprsize in parallel.c. PF.ggenterms was removed.