Fixed a bug in PF_mkDollarsParallel
The fixed bug is: if two or more $-variables are sent from a slave to
the master (or vice versa) and one of the $-variables is zero, then the
subsequent $-variables may be broken on the master, because the sender
packs d->type, d->size and d->where even if d->type == DOLZERO but the
receiver doesn't unpack those from the pack buffer.

This bug was introduced by my commit on 10 Jun. 2011.