ParFORM: rewritten code for redefine
ParFORM now shares code that is used in TFORM in order to store
information which preprocessor variables are redefined. For example,
ParFORM uses AC.numpfirstnum, AC.pfirstnum, AC.inputnumbers, etc. The
corresponding variables in ParFORM (PF.redef, PF.numredefs,
PF.mnumredefs and AC.RedefsInModuleFlag) were removed. Also PF.synchro
was removed, which was the flag that enables the mechanism that the
master broadcasts redefined preprocessor variables to the slaves (now
always this mechanism is enabled).

The code for synchronizing redefined preprocessor variables
(each slave -> master, master -> all slaves) was completely rewritten.
It now uses the 'long' packs. Thus, in practice, there is no limit in
the number/size of the variables.

PF_InitRedefinedPreVars() was renamed to PF_BroadcastRedefinedPreVars().
For now it is called at the end of module, it could be moved to the
beginning of the next module though.

The redefine statements now should work in the InParallel mode,
in principle the control of putting the variables by AC.inputnumber is
not enough when many expressions are processed concurrently though.

Added code to modify AC.pfirstnum correctly when their number is greater
than the number of the popped variable in TheUndefine(). Also added code
to allow cases like
  #define a
  #define a
  redefine a;
  #undefine a