ParFORM: implemented RHS expressions in #inside
If expression names are found in RHS in #inside, the only master
executes the contents of #inside and then broadcasts modified dollar
variables and redefined preprocessor variables. To broadcast them,
PF_BroadcastModifiedDollars() and PF_BroadcastRedefinedPreVars() are
used. To check (1) whether there are expression names in RHS, (2) which
dollar variables can be modified, and (3) which preprocessor variables
can redefined, we need to use AC.RhsExprInModuleFlag, PotModdollars and
AC.pfirstnum. In DoInside() we store the current values of them by
calling PF_StoreInsideInfo() and reset them. They are restored in
DoEndInside() by calling PF_RestoreInsideInfo().