Dollar variables when "On names"
For "On names" FORM now also prints "Dollar variables" and "Dollar
variables to be modified".

WriteLists() was made more consistent. For "On names" it doesn't print
built-in objects such as i_, iarg_, parg_, etc. They are printed for
"On allnames".

BUILTININDICES, BUILTINVECTORS and BUILTINDOLLARS, which are the numbers
of built-in objects, were added in ftypes.h.