Fixed "Nesting do and inside" bug of the forum (15 May 2012)
Checking nesting levels for Do+Inside was broken. Also for Do+Repeat.
To keep the consistency of the source code, two macros NestingChecksum()
and MesNesting() were added in declare.h. Moreover, the treatments of
AC.inexprlevel in the checksums are now fixed via NestingChecksum().

A check for missing EndInExpression at the end of modules was also
added.