ParFORM: Changes in PF_GetSlaveTimes
Now PF_GetSlaveTimes() causes actual interactions between the master and
slaves. Therefore it must be called on the master and all slaves, and it
cannot be called after PF_Terminate() is called.

Fixed the return-value to include AM.SumTime of the all slaves.

Also the prototype of GetRunningTime() has been added in declare.h.