ParFORM: synchronisation in GetRuuningTime
To keep consistency among preprocessed inputs on all processors, the
return value of GetRunningTime() is now synchronised.

Also the code for synchronising the value of termsin(expr) in pre.c was
removed because counters of expressions are synchronised at the end of
each module via PF_BroadcastExpFlags(). PF_BroadcastNumberOfTerms() was
renamed to PF_BroadcastNumber().