ParFORM: error messages at startup and fix Checkpoints
AS.printflag is set to -1 during the initialization except OpenInput(),
to print error messages occurring on slaves. The negative AS.printflag
indicates the slaves do not use the synchronized output.

The message in Terminate() was changed to show the process number.

Fixed the problem about recovery file names for Checkpoint in ParFORM
when the temporary file path is given.

Made Checkpoint in ParFORM more atomic. The file renaming is performed
after the master receives the recovery files from the all slaves.

Fixed PF_RecvFile() such that it returns -1 when it receives
PF_EMPTY_MSGTAG tag, which indicates the corresponding PF_SendFile()
was called with fd = NULL.

Some communications consisting of PF_RawSend() and PF_PF_RawRecv() were
simplified by means of PF_BroadcastNumber().

PF_CheckRepository() was modified in such a way that the master first
checks if the recovery file for the master exists and broadcasts the
result to the slaves. If it doesn't exist, return 0 immediately.