ParFORM: implementation of #pipe
The #pipe preprocessor instruction is enabled. The system command given
to #pipe is, as in #system, executed only in the master process. Then
the output of the command is broadcast to the slaves. This is
implemented as modifications in CreateStream(), CloseStream() and
ReadFromStream(). For PIPESTREAM on ParFORM, stream->buffer is used as a
send/receive buffer with its size AM.MaxStreamSize.

To broadcast the output of the command, PF_BroadcastNumber() and
PF_Bcast() are used. The declaration of the latter is added in
parallel.h.