ParFORM: fix a bug with 'maximum/minimum' $-variables
The old implementation of PF_CollectModifiedDollars() used
TwoExprCompare() for comparisons of $-variables sent by slaves. But it
can give inadequate results for numeric comparisions, if one of them is
zero. For example, for the case of
  a = { 0 }  (=> 0)
  b = { 4, 1, 1, -1 }  (=> -1)
TwoExprCompare(a, b, LESS) gives TRUE, which is inconsistent in the
sense that 0 > -1.

An example that is affected by this bug is:
  S x,y;
  L F = 1/y;
  #$xmax = -99;
  #$ymin = 0;
  if (count(x,1) > $xmax) $xmax = count_(x,1);
  if (count(y,1) < $ymin) $ymin = count_(y,1);
  ModuleOption maximum $xmax;
  ModuleOption minimum $ymin;
  .sort
  #message (xmax,ymin) = (`$xmax',`$ymin')
  .end
The old ParFORM gave (xmax,ymin) = (-99,0), which should be (0,-1).