Fix "a strange problem" reported by Kostja
http://www.nikhef.nl/~form/forum/viewtopic.php?f=8&t=141

The problem was: when a stored expression appears in the RHS,
simp6token() calls GetTable(). GetTable() may extend AS.OldNumFactors
and AS.Oldvflags if there is no enough space, but it didn't copy the
previous contents of these buffers. As a result, the information in
these buffers are lost, and actually uninitialized. Next, if another
(non-stored) expression appears in the RHS, simp6token() checks
AS.Oldvflags[n] for the expression, but it has now an uninitialized
value and may cause the "Factorized expression in RHS in an expression
of more than one term" error depending on the memory content.

Solution: make GetTable() copy the original contents in the buffers when
the reallocation occurs.