[parform] Fix a bug in disabling parallelization due to $-vars
In some cases, ParFORM failed to disable the parallelization in a module
where global $-variables are used.

For example, ParFORM showed $max=-1 for the following program:

  On allwarnings;
  S x;
  L F = <x^0>+...+<x^99>;
  #$a = 0;
  #$max = -1;
  $max = max_($max,count_(x,1));
  ModuleOption local $a;
  .sort
  #message $max=`$max'
  .end

In the above example, the parallelization is not disabled in the module
where $max is changed in a global way, but disabled for modules after
that module until the parallelization is re-enabled, e.g., by
"On parallel".