New preprocessor variable PID_
`PID_' gives the process identifier (of the master process for ParFORM).

Internally, getpid() on UNIX / GetCurrentProcessId() on Windows are
wrapped by a macro GetPID() in form3.h, which is also used for
generating temporary file names when -M option is used. Now -M works
also on Windows.

The behaviour of `PID_' for a recovered session from a checkpoint is:
the recovered session keeps using the process ID of the crushed session.