Some fixes for undefined behaviour (edited by TU)
This is extracted from the commit 29854a5b8d to synchronize and make it
easy to merge the "diagrams" branch in future. Most of them are for
removing "left shift of negative value" (e.g., -100 << 4). Note that,
on x86-64, both "a <<= 1" and "a *= 2" are naturally optimized as
"a = a + a" (at least with gcc 7.2, clang 5.0.0 and icc 17).