[ci] Fix Travis CI issues
This commit aims to resolve problems for the upgrade of the default
macOS image on Travis CI:
  https://goo.gl/H7hBsg
The upgrade brings
- "brew install mpich" fails by default (travis-ci/travis-ci#8826),
- The fixed version valgrind-3.11.0 doesn't work (Sierra is not
  supported).
Open MPI is "fragile" with Valgrind and gcov, while MPICH is not in the
APT whitelist (travis-ci/apt-package-whitelist#406), so we used to use
brewed MPICH on macOS (but the latest Valgrind gives false positives).
Now change the strategy: manually build and cache MPICH on Ubuntu and
use it with Valgrind and gcov. Test jobs on macOS are reduced to only
"osx-bin-release", which indeed checks FORM and TFORM.

It also contains refactoring of check.rb:
- Show more readable error messages,
- Add a RuboCop configuration (in my taste).