Fix polyratfun bugs due to arguments incorrectly marked clean
There are circumstances in which polyratfun arguments are passed into poly_ratfun_read,
with MUSTCLEANPRF not set. For example, if the function was inside another function argument
at the point where PolyRatFun is enabled. This causes errors because the input into the poly
routines is not properly sorted and normalized.

The issue must be caught in two places. In poly.cc poly::argument_to_poly, for univariate cases
we can check whether the order of the terms is wrong, and ensure we call normalize if so. In
the multivariate cases, normalize was always called unconditionally anyway.

In poly_ratfun_read, even though the terms are now always in the correct order, if the argument
is not marked MUSTCLEANPRF, denominators of the numerator and denominator are not dealt with.
This is resolved by unconditionally checking the minimum power of the symbols in the polynomials,
and ensuring clean = false if the minimum powers are ever negative.