Move the LogType check inside WriteStats
The updating of AC.LogHandle needs to be protected by MLOCK,MUNLOCK, to
avoid data races in tform, so move the check inside WriteStats.
We need to add a function argument to do this.

Also add defines to replace the arguments of WriteStats in its callers,
so it is clear what they mean.

Fixes #470 .