Fix memory leak in makeinteger
In case multiple functions appear in a term, and we use makeinteger on each
of them, Number allocations were not freed. Allocate the Number buffers before
looping through the term, and free after. This way the buffers are also reused.

Add a test, and force a small ulimit to ensure the crash would happen without
the fix. Skip on 32bit systems, since we ask for an 8GB ulimit.