Change EndSort type for transform statements
Since the transform statements are sorting into the workspace,
and might have to merge patches, make them use a par=1 EndSort.
This commit makes this change for all transform statements, though
only mularg is going to produce an explosion of terms here.

This fixes #230 and causes #183 to crash with a useful error
(Term too complex during normalization).