Give the SortBots their CompressBuffer
Issue #468 has two bad cases, this commit addresses one of them. It
can be when the SortBots are merging the sorted patches, that a higher
level sort in the arguments of the polyratfun creates patches and needs
to merge them. MergePatches needs a CompressBuffer, but that pointer
was 0 if we arrived there from SortBotMerge.