Don't use PutToMaster for a sLevel>0 MasterSort
During a MasterSort, if we end up calling MergePatches for a sort at
sLevel>0 (or S!=AT.S0) (when sorting PolyRatFun arguments for eg) the
output was sent to the master with PutToMaster instead of going via PutOut.
This meant the terms escaped the PolyRatFun argument and ended up at ground
level.

This fixes #468