check: add test which uses color.h
Run the color.h example as a test. Fetch color.h and put it in formlib.

Reorganise the "forcer", "checkpoint" and "multithreaded" tests into the "extra"
dir, along with the "color" test.

Run the extra tests in the valgrind and coverage runs. Some tests are disabled
under valgrind and thus also under coverage.

Fix the multithreaded test patches settings such that it runs under vorm, tvorm
and tvorm -w2 (now that they print warnings if these settings are implicitly
adjusted).