Try to create the sort file at startup.
If a bad path for FORMTMPSORT is specified, FORM doesn't notice until
it tries to create a sort file and can't. In principle one can create
a script which takes a long time to reach this point.

Try to create the file immediately at startup, and Terminate if this
fails.