Put pointer-zeroing back into SplitMerge, so GarbHand works
Commit f1b83ae removed various zeroed pointers from SplitMerge, which
does not produce wrong results, since they are "out of range" of the
sorted and merged region of the pointer array, but it leads to GarbHand
computing the total size of the terms incorrectly.

This causes valgrind errors when we write over the end of the SmallExtension,
and causes termination due to a "too small SmallExtension" which might
not actually be the case.