Reallocate large+small buffers at the end of every module
RSS decreases. If the OS is under memory pressure, it will not be
swapping out useless pages.

The SortBlocks contain pointers into the master thread's lBuffer, which
need updating after reallocation. Use an "UpdateSortBlocks" function to
do this, which is a trimmed-down version of IniSortBlocks which only
sets the pointers.