Fix termination when saving long-named expressions in a pre-proc loop
It was not intended in #500 that this warning crashes form, which happened
here when Save was called in a pre-processor loop. This is due to the
leading & of the message which triggers the source file line information.
Warning protects from the crash by temporarily setting the "iswarning"
variable which causes MesPrint to then not terminate.

This can be easily cleaned up later if Warning is refactored to take a
format string and variable arguments.