Write term size in compress buffer in GetOneTerm when reading from file and off compress
In the case if "Off Compress;" the term size was not written into the
compress buffer when reading from a file, and term compression is off.

This results in terms which are incomplete or have extra junk on the end
going through Generator; the rest is chaos.

The test is not valgrind clean for parform.