fix: various problems in poly's inverse_
The denominator of the first arg was ignored, leading to incorrect results for,
eg, inverse_(x^4-1/2,x^2). Keep it, to multiply the final result.

The input content is not removed by FORM. Remove it locally, and divide out of
the final result.

Improve the existence check of the inverse: first check the gcd of input
primitives is 1. Then, if not working in Modulus mode, attempt further
primes if an inverse is not found over the first.

Add some special cases to provide agreement with MMA's PolynomialExtendedGCD.