feature: flint support v1
This is a squash commits from https://github.com/jodavies/form/tree/flint-prf
which implement first flint support for polynomial operations.
By using `On flint;` (which is default if FORM is compiled with flint support),
flint is used for the following operations and provides a substantial
performance improvement:

PolyRatFun
FactArg
FactDollar
div_
rem_
mul_
gcd_
inverse_

It does not yet support expression factorization, or modulus mode: to be added
in the future. Also in the future the code may be refactored to make it easier
to add modulus support.